+++ WIP/Sources/FlyAwayPlus/FlyAwayPlus/Models/Place.cs
@@ -8,8 +8,8 @@ namespace FlyAwayPlus.Models
     public class Place
     {
         public int placeID { get; set; }
-        public decimal longitude { get; set; }
-        public decimal latitude { get; set; }
+        public double longitude { get; set; }
+        public double latitude { get; set; }
         public string name { get; set; }
     }
 }
